Bugfix to Lola interpreter
The grammar rule for the offset processor instantiates
two processors, but only one of them was passed to add().
This makes no difference on the resulting pipeline, unless
it is cloned, in which case the missing processor is
not cloned. Giving an event to the *cloned* pipeline would
provoke a null pointer exception.